+++ src/index.ts
@@ -1,7 +1,7 @@
 import type { API } from 'homebridge';
 
-import { PLATFORM_NAME } from './settings.js';
 import { MideaPlatform } from './platform.js';
+import { PLATFORM_NAME } from './settings.js';
 
 /**
  * This method registers the platform with Homebridge